Refactor MainHeader
