+++ client/src/redux/refProfile/selectors.ts
@@ -2,6 +2,6 @@ import { prefix } from './actions';
 import { ReduxState } from '../../types';
 
 export const loading = (state: ReduxState) => state[prefix].loading;
-export const data = (state: ReduxState) => state[prefix].data;
+export const refProfileInfo = (state: ReduxState) => state[prefix].data;
 export const params = (state: ReduxState) => state[prefix].params;
 export const refLinkError = (state: ReduxState) => state[prefix].refLinkError;
